+++ src/tests/CMakeLists.txt
@@ -11,6 +11,7 @@ set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
 FetchContent_MakeAvailable(googletest)
 
 add_executable(unit_tests unit_tests.cpp)
+add_executable(robot_tests robot_tests.cpp)
 
 target_link_libraries(unit_tests
     algorithm_lib
@@ -18,5 +19,11 @@ target_link_libraries(unit_tests
     GTest::gtest_main
 )
 
+target_link_libraries(robot_tests
+    algorithm_lib
+    robot_lib
+    GTest::gtest_main
+)
+
 include(GoogleTest)
-gtest_discover_tests(unit_tests)
+gtest_discover_tests(unit_tests robot_tests)
