+++ src/tests/input.txt
@@ -1,0 +1,4 @@
+5 20
+-2 -1 0 2 3
+-2 -2 1 2 -2
+-2 -2 5 9 9
